% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_aum.R
\name{table_aum}
\alias{table_aum}
\title{Make a Frequency table including a row for NAs}
\usage{
table_aum(x)
}
\arguments{
\item{x}{factor}
}
\value{
A tibble
}
\description{
Make a Frequency table including a row for NAs
}
\examples{
table_aum(iris$Species)
}
